refactor: dissolve the importing domain module
Importing was the last slice with a domain module, so its own
vocabulary still lived in two places. Moving the pure code into
services closes that gap. With no domain left anywhere, the import
checker and the docs can drop the rules that only existed while
the move was going on.

The document rejection types now sit next to the reader that makes
them, instead of travelling with the scan types.
