+++ mpvqc/importing/services/__init__.py
@@ -2,8 +2,45 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from .concerns import LoadFoundVideo as LoadFoundVideo
+from .concerns import SessionConcern as SessionConcern
+from .concerns import SessionMerge as SessionMerge
+from .concerns import SessionReplace as SessionReplace
+from .concerns import SessionResolved as SessionResolved
+from .concerns import SessionUnresolved as SessionUnresolved
+from .concerns import SubtitlesConcern as SubtitlesConcern
+from .concerns import SubtitlesLoad as SubtitlesLoad
+from .concerns import SubtitlesResolved as SubtitlesResolved
+from .concerns import SubtitlesSkip as SubtitlesSkip
+from .concerns import SubtitlesUnresolved as SubtitlesUnresolved
+from .concerns import VideoConcern as VideoConcern
+from .concerns import VideoLoad as VideoLoad
+from .concerns import VideoResolved as VideoResolved
+from .concerns import VideoSkip as VideoSkip
+from .concerns import VideoUnresolved as VideoUnresolved
+from .documents import ParsedDocument as ParsedDocument
+from .documents import parse_classic as parse_classic
+from .documents import parse_v1 as parse_v1
+from .errors import ErrorsAbsent as ErrorsAbsent
+from .errors import ErrorsPresent as ErrorsPresent
+from .errors import ImportErrors as ImportErrors
+from .file_kind import DOCUMENT_EXTENSIONS as DOCUMENT_EXTENSIONS
+from .file_kind import SUBTITLE_EXTENSIONS as SUBTITLE_EXTENSIONS
+from .file_kind import ClassifiedPaths as ClassifiedPaths
+from .file_kind import classify_paths as classify_paths
 from .importer import ImportService as ImportService
+from .pending import PendingImport as PendingImport
+from .plan import FinishedPlan as FinishedPlan
+from .plan import NotAsked as NotAsked
+from .plan import UnfinishedPlan as UnfinishedPlan
+from .plan import finish_plan as finish_plan
+from .plan import make_plan as make_plan
+from .reader import DocumentRejectionReason as DocumentRejectionReason
+from .reader import RejectedDocument as RejectedDocument
 from .reader import read_documents as read_documents
+from .scan import ScanResult as ScanResult
+from .scan import SubtitleSource as SubtitleSource
+from .scan import VideoSource as VideoSource
 from .scan import scan as scan
 from .settings import ImportSettingsService as ImportSettingsService
 from .subtitle_videos import find_videos_in_subtitles as find_videos_in_subtitles
