+++ mpvqc/importing/viewmodels/settings_dialog.py
@@ -6,8 +6,7 @@ import inject
 from PySide6.QtCore import Property, QCoreApplication, QObject, Signal, Slot
 from PySide6.QtQml import QmlElement
 
-from mpvqc.importing.domain import LoadFoundVideo
-from mpvqc.importing.services import ImportSettingsService
+from mpvqc.importing.services import ImportSettingsService, LoadFoundVideo
 
 QML_IMPORT_NAME = "io.github.mpvqc.mpvQC.Python"
 QML_IMPORT_MAJOR_VERSION = 1
