+++ pyproject.toml
@@ -173,14 +173,6 @@ files = [
     "mpvqc/exporting/viewmodels/file_dialog.py",
     "mpvqc/exporting/wiring.py",
     "mpvqc/importing/__init__.py",
-    "mpvqc/importing/domain/__init__.py",
-    "mpvqc/importing/domain/concerns.py",
-    "mpvqc/importing/domain/documents.py",
-    "mpvqc/importing/domain/errors.py",
-    "mpvqc/importing/domain/file_kind.py",
-    "mpvqc/importing/domain/pending.py",
-    "mpvqc/importing/domain/plan.py",
-    "mpvqc/importing/domain/scan.py",
     "mpvqc/importing/enums/__init__.py",
     "mpvqc/importing/enums/session_mode.py",
     "mpvqc/importing/enums/step_kind.py",
@@ -189,7 +181,13 @@ files = [
     "mpvqc/importing/models/subtitles.py",
     "mpvqc/importing/models/videos.py",
     "mpvqc/importing/services/__init__.py",
+    "mpvqc/importing/services/concerns.py",
+    "mpvqc/importing/services/documents.py",
+    "mpvqc/importing/services/errors.py",
+    "mpvqc/importing/services/file_kind.py",
     "mpvqc/importing/services/importer.py",
+    "mpvqc/importing/services/pending.py",
+    "mpvqc/importing/services/plan.py",
     "mpvqc/importing/services/reader.py",
     "mpvqc/importing/services/scan.py",
     "mpvqc/importing/services/settings.py",
