+++ test/importing/services/test_make_plan.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, NamedTuple
 
 import pytest
 
-from mpvqc.importing.domain import (
+from mpvqc.importing.services import (
     DocumentRejectionReason,
     ErrorsAbsent,
     ErrorsPresent,
