+++ test/importing/services/test_reader.py
@@ -11,8 +11,7 @@ from typing import NamedTuple
 
 import pytest
 
-from mpvqc.importing.domain import DocumentRejectionReason, RejectedDocument
-from mpvqc.importing.services import read_documents
+from mpvqc.importing.services import DocumentRejectionReason, RejectedDocument, read_documents
 
 DOCUMENT_FORMAT_README = Path(__file__).parents[3] / "docs" / "document-format" / "README.md"
